requirements.txt
# Core dependencies
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
jinja2>=3.1.2
python-multipart

# Vector search and embedding (ChromaDB)
chromadb==0.4.24
numpy==1.26.4

# Document parsing
python-docx==1.1.0
PyMuPDF==1.23.25

# YAML for config loading
pyyaml==6.0.1